Get rid of dead functions in MangaWrapper, rename it to ImageDownloader
